fix: repoint stale vi.mock path and restore home href after trace move
