+++ web/src/features/traces/components/ObservationDetailView/ObservationDetailView.tsx
@@ -55,7 +55,7 @@ import { useCommentedPaths } from "@/src/features/comments/hooks/useCommentedPat
 import { api } from "@/src/utils/api";
 
 // Extracted components
-import { ObservationDetailViewHeader } from "./ObservationDetailViewHeader";
+import { ObservationDetailViewHeader } from "./components/ObservationDetailViewHeader";
 import { TraceLogView } from "../TraceLogView/TraceLogView";
 import { TRACE_VIEW_CONFIG } from "@/src/features/traces/config/trace-view-config";
 import { useV4Beta } from "@/src/features/events/hooks/useV4Beta";
