+++ web/src/features/traces/components/ObservationDetailView/components/ObservationMetadataBadgeModel.tsx
