+++ web/src/features/traces/components/ObservationDetailView/components/ObservationMetadataBadgeModelParameters.tsx
