+++ web/src/features/traces/components/ObservationMetadataBadgesSimple/ObservationMetadataBadgesSimple.clienttest.tsx
