+++ web/src/features/traces/components/ObservationMetadataBadgesSimple/ObservationMetadataBadgesSimple.tsx
