+++ web/src/features/traces/components/ObservationMetadataBadgesTooltip.tsx
