+++ web/src/features/traces/components/TraceMetadataBadges.tsx
